添加会话、删除会话，bug fixed

+++ src/main/java/com/example/echatbackend/controller/MessageController.java
@@ -31,16 +31,22 @@ public class MessageController extends BaseController {
     // 删除信息
     @PostMapping("/mes/removeMessage")
     public void removeMessage(@NotNull @RequestBody JSONObject request) {
-        int id = request.getInteger("id");
+        Integer id = request.getInteger("id");
+        if (id == null) {
+            return;
+        }
         messageService.deleteMessage(id);
     }
 
     // 加载更多消息
     @PostMapping("/mes/loadMoreMessages")
     public ResponseEntity<Object> loadMoreMessages(@NotNull @RequestBody JSONObject request) {
-        int roomId = request.getInteger("roomId");
-        int offset = request.getInteger("offset");
-        int limit = request.getInteger("limit");
+        Integer roomId = request.getInteger("roomId");
+        Integer offset = request.getInteger("offset");
+        Integer limit = request.getInteger("limit");
+        if (roomId == null || offset == null || limit == null) {
+            return requestFail(-1, "参数错误");
+        }
         Page<Message> messages = messageService.getMoreMessage(roomId, offset, limit);
         List<Message> messageList = messages.getContent();
         return ResponseEntity.ok(JSONArray.parseArray(JSON.toJSONString(messageList.stream().map(Message::show).toArray(JSONObject[]::new))));
